@@ -4,13 +4,13 @@ version: 1.1.0
 homepage: http://laomengit.com/plugin/upgrade.html
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  dio: ^3.0.9
+  dio: ^4.0.0
 
 dev_dependencies:
   flutter_test:
