@@ -12,19 +12,19 @@ class LiquidLinearProgressIndicator extends ProgressIndicator {
   final double borderRadius;
 
   ///The widget to show in the center of the progress indicator.
-  final Widget center;
+  final Widget? center;
 
   ///The direction the liquid travels.
   final Axis direction;
 
   LiquidLinearProgressIndicator({
-    Key key,
+    Key? key,
     double value = 0.5,
-    Color backgroundColor,
-    Animation<Color> valueColor,
-    this.borderWidth,
-    this.borderColor,
-    this.borderRadius,
+    Color? backgroundColor,
+    Animation<Color>? valueColor,
+    required this.borderWidth,
+    required this.borderColor,
+    required this.borderRadius,
     this.center,
     this.direction = Axis.horizontal,
   }) : super(
@@ -32,12 +32,7 @@ class LiquidLinearProgressIndicator extends ProgressIndicator {
           value: value,
           backgroundColor: backgroundColor,
           valueColor: valueColor,
-        ) {
-    if (borderWidth != null && borderColor == null ||
-        borderColor != null && borderWidth == null) {
-      throw ArgumentError("borderWidth and borderColor should both be set.");
-    }
-  }
+        );
 
   Color _getBackgroundColor(BuildContext context) =>
       backgroundColor ?? Color(0x0000BFFF); //Theme.of(context).backgroundColor;
@@ -49,8 +44,7 @@ class LiquidLinearProgressIndicator extends ProgressIndicator {
   State<StatefulWidget> createState() => _LiquidLinearProgressIndicatorState();
 }
 
-class _LiquidLinearProgressIndicatorState
-    extends State<LiquidLinearProgressIndicator> {
+class _LiquidLinearProgressIndicatorState extends State<LiquidLinearProgressIndicator> {
   @override
   Widget build(BuildContext context) {
     return ClipPath(
@@ -70,7 +64,7 @@ class _LiquidLinearProgressIndicatorState
         child: Stack(
           children: <Widget>[
             Wave(
-              value: widget.value,
+              value: widget.value!,
               color: widget._getValueColor(context),
               direction: widget.direction,
             ),
@@ -86,7 +80,7 @@ class _LinearPainter extends CustomPainter {
   final Color color;
   final double radius;
 
-  _LinearPainter({@required this.color, @required this.radius});
+  _LinearPainter({required this.color, required this.radius});
 
   @override
   void paint(Canvas canvas, Size size) {
@@ -109,9 +103,9 @@ class _LinearBorderPainter extends CustomPainter {
   final double radius;
 
   _LinearBorderPainter({
-    @required this.color,
-    @required this.width,
-    @required this.radius,
+    required this.color,
+    required this.width,
+    required this.radius,
   });
 
   @override
@@ -127,8 +121,7 @@ class _LinearBorderPainter extends CustomPainter {
     final alteredRadius = radius ?? 0;
     canvas.drawRRect(
         RRect.fromRectAndRadius(
-          Rect.fromLTWH(
-              width / 2, width / 2, size.width - width, size.height - width),
+          Rect.fromLTWH(width / 2, width / 2, size.width - width, size.height - width),
           Radius.circular(alteredRadius - width ?? 0),
         ),
         paint);
@@ -144,7 +137,7 @@ class _LinearBorderPainter extends CustomPainter {
 class _LinearClipper extends CustomClipper<Path> {
   final double radius;
 
-  _LinearClipper({@required this.radius});
+  _LinearClipper({required this.radius});
 
   @override
   Path getClip(Size size) {
