@@ -12,9 +12,9 @@ import 'liquid_progress_indicator.dart';
 ///
 class SimpleAppUpgradeWidget extends StatefulWidget {
   const SimpleAppUpgradeWidget(
-      {@required this.title,
+      {required this.title,
       this.titleStyle,
-      @required this.contents,
+      required this.contents,
       this.contentStyle,
       this.cancelText,
       this.cancelTextStyle,
@@ -41,7 +41,7 @@ class SimpleAppUpgradeWidget extends StatefulWidget {
   ///
   /// 标题样式
   ///
-  final TextStyle titleStyle;
+  final TextStyle? titleStyle;
 
   ///
   /// 升级提示内容
@@ -51,47 +51,47 @@ class SimpleAppUpgradeWidget extends StatefulWidget {
   ///
   /// 提示内容样式
   ///
-  final TextStyle contentStyle;
+  final TextStyle? contentStyle;
 
   ///
   /// 下载进度条
   ///
-  final Widget progressBar;
+  final Widget? progressBar;
 
   ///
   /// 进度条颜色
   ///
-  final Color progressBarColor;
+  final Color? progressBarColor;
 
   ///
   /// 确认控件
   ///
-  final String okText;
+  final String? okText;
 
   ///
   /// 确认控件样式
   ///
-  final TextStyle okTextStyle;
+  final TextStyle? okTextStyle;
 
   ///
   /// 确认控件背景颜色,2种颜色左到右线性渐变
   ///
-  final List<Color> okBackgroundColors;
+  final List<Color>? okBackgroundColors;
 
   ///
   /// 取消控件
   ///
-  final String cancelText;
+  final String? cancelText;
 
   ///
   /// 取消控件样式
   ///
-  final TextStyle cancelTextStyle;
+  final TextStyle? cancelTextStyle;
 
   ///
   /// app安装包下载url,没有下载跳转到应用宝等渠道更新
   ///
-  final String downloadUrl;
+  final String? downloadUrl;
 
   ///
   /// 圆角半径
@@ -106,18 +106,18 @@ class SimpleAppUpgradeWidget extends StatefulWidget {
   ///
   /// ios app id,用于跳转app store
   ///
-  final String iosAppId;
+  final String? iosAppId;
 
   ///
   /// 指定跳转的应用市场，
   /// 如果不指定将会弹出提示框，让用户选择哪一个应用市场。
   ///
-  final AppMarketInfo appMarketInfo;
+  final AppMarketInfo? appMarketInfo;
 
-  final VoidCallback onCancel;
-  final VoidCallback onOk;
-  final DownloadProgressCallback downloadProgress;
-  final DownloadStatusChangeCallback downloadStatusChange;
+  final VoidCallback? onCancel;
+  final VoidCallback? onOk;
+  final DownloadProgressCallback? downloadProgress;
+  final DownloadStatusChangeCallback? downloadStatusChange;
 
   @override
   State<StatefulWidget> createState() => _SimpleAppUpgradeWidget();
@@ -227,11 +227,11 @@ class _SimpleAppUpgradeWidget extends State<SimpleAppUpgradeWidget> {
   _buildCancelActionButton() {
     return Ink(
       decoration: BoxDecoration(
-          borderRadius: BorderRadius.only(
-              bottomLeft: Radius.circular(widget.borderRadius))),
+          borderRadius:
+              BorderRadius.only(bottomLeft: Radius.circular(widget.borderRadius))),
       child: InkWell(
-          borderRadius: BorderRadius.only(
-              bottomLeft: Radius.circular(widget.borderRadius)),
+          borderRadius:
+              BorderRadius.only(bottomLeft: Radius.circular(widget.borderRadius)),
           child: Container(
             height: 45,
             alignment: Alignment.center,
@@ -257,8 +257,7 @@ class _SimpleAppUpgradeWidget extends State<SimpleAppUpgradeWidget> {
           bottomLeft: Radius.circular(widget.borderRadius));
     }
     var _okBackgroundColors = widget.okBackgroundColors;
-    if (widget.okBackgroundColors == null ||
-        widget.okBackgroundColors.length != 2) {
+    if (widget.okBackgroundColors == null || widget.okBackgroundColors!.length != 2) {
       _okBackgroundColors = [
         Theme.of(context).primaryColor,
         Theme.of(context).primaryColor
@@ -269,7 +268,7 @@ class _SimpleAppUpgradeWidget extends State<SimpleAppUpgradeWidget> {
           gradient: LinearGradient(
               begin: Alignment.topLeft,
               end: Alignment.bottomRight,
-              colors: [_okBackgroundColors[0], _okBackgroundColors[1]]),
+              colors: [_okBackgroundColors![0], _okBackgroundColors[1]]),
           borderRadius: borderRadius),
       child: InkWell(
         borderRadius: borderRadius,
@@ -294,9 +293,11 @@ class _SimpleAppUpgradeWidget extends State<SimpleAppUpgradeWidget> {
         LiquidLinearProgressIndicator(
           value: _downloadProgress,
           direction: Axis.vertical,
-          valueColor: AlwaysStoppedAnimation(widget.progressBarColor ??
-              Theme.of(context).primaryColor.withOpacity(0.4)),
+          valueColor: AlwaysStoppedAnimation(
+              widget.progressBarColor ?? Theme.of(context).primaryColor.withOpacity(0.4)),
           borderRadius: widget.borderRadius,
+          borderWidth: 1,
+          borderColor: Colors.red,
         );
   }
 
@@ -305,18 +306,18 @@ class _SimpleAppUpgradeWidget extends State<SimpleAppUpgradeWidget> {
   ///
   _clickOk() async {
     widget.onOk?.call();
-    if (Platform.isIOS) {
+    if (Platform.isIOS && widget.iosAppId != null) {
       //ios 需要跳转到app store更新，原生实现
-      FlutterUpgrade.toAppStore(widget.iosAppId);
+      FlutterUpgrade.toAppStore(widget.iosAppId!);
       return;
     }
-    if (widget.downloadUrl == null || widget.downloadUrl.isEmpty) {
+    if (widget.downloadUrl == null || widget.downloadUrl!.isEmpty) {
       //没有下载地址，跳转到第三方渠道更新，原生实现
       FlutterUpgrade.toMarket(appMarketInfo: widget.appMarketInfo);
       return;
     }
     String path = await FlutterUpgrade.apkDownloadPath;
-    _downloadApk(widget.downloadUrl, '$path/$_downloadApkName');
+    _downloadApk(widget.downloadUrl!, '$path/$_downloadApkName');
   }
 
   ///
@@ -351,7 +352,7 @@ class _SimpleAppUpgradeWidget extends State<SimpleAppUpgradeWidget> {
     } catch (e) {
       print('$e');
       _downloadProgress = 0;
-      _updateDownloadStatus(DownloadStatus.error,error: e);
+      _updateDownloadStatus(DownloadStatus.error, error: e);
     }
   }
 
